web/django_day11/django_day12_assignment/db_ws_8_5/requirements.txt
asgiref==3.3.4
attrs==23.1.0
Django==3.2
django-seed==0.2.2
djangorestframework==3.12.4
drf-spectacular==0.26.5
Faker==8.1.1
inflection==0.5.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.10.6
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
uritemplate==4.1.1
